requirements.txt
amqp
arrow
celery
cookiecutter
Django
django-crontab
django-daterange-filter
django-debug-toolbar
django-extensions
django-import-export
django-preventconcurrentlogins
djangorestframework
djangorestframework-expiring-authtoken
Fabric
factory-boy
Faker
ipdb
ipython
psycopg2
python-dateutil
python-magic
pytz
PyYAML
requests
tablib
vine
vpnc
wcwidth
websocket-client
Werkzeug
xlrd
XlsxWriter
django-request-id
django-user-sessions
django-two-factor-auth
python-magic
django-celery-beat
django-test-plus
coverage
mock
django-environ
